+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [[#188]](https://github.com/Marcin-Migdal/m-component-library/issues/188) **[PATCH]** Fix overlay disabled key press, add onClose callback that is called on "esc" button press.
 - [[#189]](https://github.com/Marcin-Migdal/m-component-library/issues/189) **[PATCH]** Controlled value of the ColorPicker cannot be null, now for controlled input value has to be clearly and properly defined.
 - [[#192]](https://github.com/Marcin-Migdal/m-component-library/issues/192) **[PATCH]** Touched value returned from useForm is now cleared when using resetForm function.
+- [[#190]](https://github.com/Marcin-Migdal/m-component-library/issues/190) **[PATCH]** Add hideOnDisabledOptions props to DropdownMenu component.
 
 ### Release date 20.03.2025 - Release 2.1.1